backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
supabase==2.0.3
openai==0.28.1
psycopg2-binary==2.9.9
